💥 Move NestJS app logic and remove app factory option
