--- src/nestjs/factory/index.ts
@@ -1,1 +1,0 @@
-export * from './app-factory.js';
