+++ src/nestjs/healthcheck/module.spec.ts
@@ -4,8 +4,8 @@ import 'jest-extended';
 import supertest from 'supertest';
 import TestAgent from 'supertest/lib/agent.js';
 import { getLoggedErrors, spyOnLogger } from '../../testing.js';
+import { createApp } from '../app/index.js';
 import { AuthModule } from '../auth/index.js';
-import { createApp } from '../factory/index.js';
 import { LoggerModule } from '../logging/index.js';
 import { generateOpenApiDocument } from '../openapi/utils.test.js';
 import { BaseHealthIndicatorService } from './base-health-indicator.service.js';
