+++ src/nestjs/index.ts
@@ -1,7 +1,7 @@
+export * from './app/index.js';
 export * from './auth/index.js';
 export * from './errors/index.js';
 export * from './events/index.js';
-export * from './factory/index.js';
 export * from './healthcheck/index.js';
 export * from './logging/index.js';
 export * from './openapi/index.js';
